+++ riscv/insts/integer/base/i/csrrwi.kt
@@ -20,8 +20,8 @@ val csrrwi = Instruction(
         parser = CSRTypeParser,
         impl16 = NoImplementation,
         impl32 = RawImplementation { mcode, sim ->
-            val imm = mcode[InstructionField.RS1].toInt()
-            val vcsr = sim.getReg(32).toInt()
+            val imm = mcode[InstructionField.RS1].toLong()
+            val vcsr = sim.getReg(32).toLong()
             sim.setReg(mcode[InstructionField.RD].toInt(), vcsr)
             sim.setReg(32, imm)
             sim.incrementPC(mcode.length)
